\name{stock_data}
\alias{RSHCQ}
\alias{}
\docType{data}
\title{}
\description{
  Data set consisting of 69 columns of data about the German elections
  2002 and 2005.

  TODO: all
}
\usage{
data("")
data("")
}
\format{
  TODO
}
\source{
  The data was obtained from TODO on 2016-01-03.
}
\author{Marius Hofert}
\examples{
data("")
data("")
}
\keyword{datasets}